Enemy and Player changes.
Implemented the enemy movement. Will move randomly between any amount of
points you give it. It is all modular and so points can be removed and
added and the enemy should adapt accordingly. Created the intial player
script, doesn't control any movement but the player has a certain amount
of lives and upon collision with an enemy, will lose one of those lives
(or more than 1 if a higher value is  assigned to an enemy). When an
enemy hits a player, it will also play a sound that simulates the player
being hit. Finally via a Death() function, when the player has lost
their maximum amount of lives, a message is outputted to the console to
indicate as such and a boolean isDead is set to true. Someone else who
handles what happens after their death (I think Tom) will act according
to the value of isDead and so on.