Made game into around 5 minutes of playable gameplay.
Game can now be beaten, and will take around 3-4 mins. Got rid of most
bugs, obviously game is still shit. Still waiting on the level art and
sounds from ryan, at the very least the game is in a playable state and
can be handed in like this and still count if absolutely needed.

gonna kms now.